+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Chat-App
 
 
+# 2.10.1
+ * Update server config.
+ * Improved webpack dev/prod scenarios.
+ * Minor fixes.
+
 # 2.9.1
  * Added Websockets notification once some user is leave the chat.
  * Minor style improvements.